pyproject.toml
[tool.poetry]
name = "movies-depi-grad-project"
version = "0.1.0"
description = ""
authors = ["Morsi <morsim33@gmail.com>", "Salah <salahalgamasy@gmail.com>", "Shahd <shahd.hamdi162@gmail.com>", "Mohamed <mohamedelgohary978@gmail.com>", "Alyaa <Alyaamahmoodx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
mlflow = "^2.17.0"
streamlit = "^1.39.0"
scikit-learn = "^1.5.2"
faiss-cpu = "^1.9.0"
sentence-transformers = "^3.2.0"
pyodbc = "^5.2.0"
sqlalchemy = "^2.0.36"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
